requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.4.17        # via requests
chardet==3.0.4            # via requests
click==6.7                # via flask, python-dotenv
flask==0.12.2
idna==2.5                 # via requests
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
markupsafe==1.0           # via jinja2
python-dotenv==0.6.4
requests==2.18.1
urllib3==1.21.1           # via requests
werkzeug==0.12.2          # via flask
